Use `Connected` to replace `Endpoint`
