+++ kernel/aster-nix/src/net/socket/unix/stream/mod.rs
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
 mod connected;
-mod endpoint;
 mod init;
 mod listener;
 mod socket;
